+++ src/clj_kondo/impl/types.clj
@@ -36,6 +36,7 @@
     :map
     :nil
     :set
+    :sorted-set
     :fn
     :keyword
     :symbol
@@ -52,7 +53,8 @@
     :any
     :float
     :var
-    :ilookup})
+    :ilookup
+    :array})
 
 (def built-in-specs
   {'clojure.core clojure-core
@@ -75,6 +77,7 @@
    :nil #{:seqable}
    :coll #{:seqable}
    :set #{:seqable :coll :ifn :ilookup}
+   :sorted-set #{:set :seqable :coll :ifn :ilookup}
    :fn #{:ifn}
    :keyword #{:ifn}
    :symbol #{:ifn}
@@ -85,29 +88,32 @@
    :sequential #{:coll :seqable}
    :sorted-map #{:map :seqable :associative :coll :ifn :ilookup}
    :atom #{:ideref}
-   :var #{:ideref :ifn}})
+   :var #{:ideref :ifn}
+   :array #{:seqable :ilookup}})
 
 (def could-be-relations
   {:char-sequence #{:string}
    :int #{:neg-int :nat-int :pos-int}
    :number #{:neg-int :pos-int :nat-int :int :double :byte :ratio}
-   :coll #{:map :sorted-map :vector :set :list  :associative :seq :sequential :ifn :stack
-           :ilookup}
-   :seqable #{:coll :vector :set :map :associative
+   :coll #{:map :sorted-map :vector :set :sorted-set :list :associative :seq
+           :sequential :ifn :stack :ilookup}
+   :seqable #{:coll :vector :set :sorted-set :map :associative
               :char-sequence :string :nil
-              :list :seq :sequential :ifn :stack :sorted-map :ilookup}
+              :list :seq :sequential :ifn :stack :sorted-map :ilookup :array}
    :associative #{:map :vector :sequential :stack :sorted-map}
-   :ifn #{:fn :transducer :symbol :keyword :map :set :vector :associative :seqable :coll
-          :sequential :stack :sorted-map :var :ideref :ilookup}
+   :ifn #{:fn :transducer :symbol :keyword :map :set :sorted-set :vector
+          :associative :seqable :coll :sequential :stack :sorted-map :var
+          :ideref :ilookup}
    :fn #{:transducer}
    :nat-int #{:pos-int}
    :seq #{:list :stack}
    :stack #{:list :vector :seq :sequential :seqable :coll :ifn :associative :ilookup}
    :sequential #{:seq :list :vector :ifn :associative :stack :ilookup}
    :map #{:sorted-map}
+   :set #{:sorted-set}
    :ideref #{:atom :var :ifn}
-   :ilookup #{:map :set :sorted-map :coll :seqable :ifn :associative :vector
-              :sequential :stack}})
+   :ilookup #{:map :set :sorted-set :sorted-map :coll :seqable :ifn :associative
+              :vector :sequential :stack :array}})
 
 (def misc-types #{:boolean :atom :regex :char})
 
@@ -152,12 +158,14 @@
    :transducer "transducer"
    :seqable-or-transducer "seqable or transducer"
    :set "set"
+   :sorted-set "sorted set"
    :char-sequence "char sequence"
    :sequential "sequential collection"
    :throwable "throwable"
    :sorted-map "sorted map"
    :var "var"
-   :ilookup "ILookup"})
+   :ilookup "ILookup"
+   :array "array"})
 
 (defn label [k]
   (cond