+++ src/clj_kondo/impl/types/clojure/core.clj
@@ -143,7 +143,9 @@
    ;; 272
    'butlast seqable->nilable-seq
    ;; 283 'defn
-   ;; 338 'to-array
+   ;; 338 
+   'to-array {:arities {1 {:args [:nilable/coll]
+                           :ret :array}}}
    ;; 346 'cast
    ;; 353
    'vector {:arities {:varargs {:ret :vector}}}
@@ -156,9 +158,16 @@
    'hash-set {:arities {:varargs {:ret :set}}}
    ;; 398
    'sorted-map {:arities {:varargs {:ret :sorted-map}}}
-   ;; 407 'sorted-map-by
-   ;; 417 'sorted-set
-   ;; 425 'sorted-set-by
+   ;; 407 
+   'sorted-map-by {:arities {:varargs {:args [:ifn {:op :rest
+                                                    :spec [:any :any]}]
+                                       :ret :sorted-map}}}
+   ;; 417 
+   'sorted-set {:arities {:varargs {:ret :sorted-set}}}
+   ;; 425 
+   'sorted-set-by {:arities {:varargs {:args [:ifn {:op :rest
+                                                    :spec :any}]
+                                       :ret :sorted-set}}}
    ;; 436
    'nil? any->boolean
    ;; 444 'defmacro
@@ -726,10 +735,22 @@
    'doto {:fn first}
    ;; 3871 'memfn
    ;; 3884 'time
-   ;; 3898 'alength
-   ;; 3905 'aclone
-   ;; 3912 'aget
-   ;; 3923 'aset
+   ;; 3898
+   'alength {:arities {1 {:args [:array] :ret :int}}}
+   ;; 3905 
+   'aclone {:arities {1 {:args [:array] :ret :array}}}
+   ;; 3912 
+   'aget {:arities {2 {:args [:array :any] :ret :any}
+                    3 {:args [:array :any :any] :ret :any}
+                    :varargs {:min-arity 2 
+                              :args [:array :any {:op :rest :spec :any}]
+                              :ret :any}}}
+   ;; 3923
+   'aset {:arities {3 {:args [:array :any :any] :ret :any}
+                    4 {:args [:array :any :any :any] :ret :any}
+                    :varargs {:min-arity 3
+                              :args [:array :any {:op :rest :spec :any} :any]
+                              :ret :any}}}
    ;; 3986 'make-array
    ;; 4003 'to-array-2d
    ;; 4018 'macroexpand-1